investment-analysis-tool/requirements.txt
pandas
numpy
jinja2
flask
markdown
plaid-python
python-dotenv
